refactor: extract ALE host Codex CLI commands
