+++ loopx/cli_commands/__init__.py
@@ -16,6 +16,10 @@ from .agents_last_exam_baked_input import (
     handle_agents_last_exam_baked_input_command,
     register_agents_last_exam_baked_input_commands,
 )
+from .agents_last_exam_host_codex import (
+    handle_agents_last_exam_host_codex_command,
+    register_agents_last_exam_host_codex_commands,
+)
 from .agents_last_exam_local_plan import (
     handle_agents_last_exam_local_plan_command,
     register_agents_last_exam_local_plan_commands,
@@ -136,6 +140,7 @@ from .worker_bridge import handle_worker_bridge_command, register_worker_bridge_
 __all__ = [
     "handle_agents_last_exam_command",
     "handle_agents_last_exam_baked_input_command",
+    "handle_agents_last_exam_host_codex_command",
     "handle_agents_last_exam_launch_dry_run_command",
     "handle_agents_last_exam_local_plan_command",
     "handle_agents_last_exam_runner_source_command",
@@ -188,6 +193,7 @@ __all__ = [
     "handle_worker_bridge_command",
     "register_agents_last_exam_commands",
     "register_agents_last_exam_baked_input_commands",
+    "register_agents_last_exam_host_codex_commands",
     "register_agents_last_exam_launch_dry_run_commands",
     "register_agents_last_exam_local_plan_commands",
     "register_agents_last_exam_runner_source_commands",
